Fix git conflict in Darwin OS
1. rename file `07_dijkstras_algorithm/golang/01_dijkstras_algorithm.go` to
  `07_dijkstras_algorithm/Golang/01_dijkstras_algorithm-alias.go`,
2. delete folder `golang`

Git warning:
```
warning: the following paths have collided (e.g. case-sensitive paths
on a case-insensitive filesystem) and only one from the same
colliding group is in the working tree:

  '07_dijkstras_algorithm/Golang/01_dijkstras_algorithm.go'
  '07_dijkstras_algorithm/golang/01_dijkstras_algorithm.go'
```
